feat: Item API, Pagination Layout

+++ app/Http/Controllers/UomController.php
@@ -7,6 +7,7 @@ use App\Http\Requests\UpdateUomRequest;
 use App\Models\Uom;
 use Illuminate\Http\Request;
 use Inertia\Inertia;
+use Illuminate\Support\Facades\DB;
 
 class UomController extends Controller
 {
@@ -77,7 +78,19 @@ class UomController extends Controller
      */
     public function destroy(Uom $uom)
     {
-        //
+
+$itemUsingUom = DB::table('item_uoms')
+        ->join('items', 'item_uoms.item_id', '=', 'items.id')
+        ->where('item_uoms.uom_id', $uom->id)
+        ->whereNull('items.deleted_at')
+        ->select('items.name') 
+        ->first();             
+
+    if ($itemUsingUom) {
+        $errorMessage = "UOM ini tidak dapat dihapus karena sedang digunakan oleh Barang: {$itemUsingUom->name}";
+
+        return redirect()->back()->withErrors(['msg' => $errorMessage]);
+    }
         $uom->delete();
 
         return redirect()->route('uoms.index')
